backend/requirements.txt
flask
flask-cors
langchain
langchain-community
langchain_openai
langchain_pinecone
openai
pinecone
python-dotenv
pypdf
python-docx
pandas
sqlite-utils